modification batch import région + gestion de la publication d'aides (#9)
* Mise à jour des batchs d'import
* Gestion de la publication (et dépublication d'une aide)

+++ database/migrations/2018_01_25_111158_drop_unused_cols_calls_for_projects.php
@@ -13,9 +13,12 @@ class DropUnusedColsCallsForProjects extends Migration
      */
     public function up()
     {
-        Schema::table('calls_for_projects', function (Blueprint $table) {
-            $table->dropColumn(['project_holder_id', 'beneficiary_id', 'perimeter_id']);
-        });
+        if (!Schema::hasTable('calls_for_projects')) {
+            Schema::table('calls_for_projects', function (Blueprint $table) {
+                $table->dropColumn(['project_holder_id', 'beneficiary_id', 'perimeter_id']);
+            });
+        }
+       
     }
 
     /**
